Complete consuming Python async close lifecycle
