+++ plans/reviews/active/python-interop-m7-wave6-async-close-design-review-round1.md
@@ -1,0 +1,30 @@
+I've read the M7 Wave 6 plan, the current sealed-identity/foreign-object machinery, the async runtime/declaration path, the codegen (`python_interop_async.rs`, `python_interop_direct.rs`, `class_method_emitter.rs`), the HIR-level class validation (`class_body_lowering.rs`), gate discipline (`python_interop.rs`, coroutine contract tests), and the two architecture docs. Below is the design review.
+
+## What the proposal gets right
+
+- **Atomic Open→Closing transition.** Producing the close task's lease under the same mutex acquisition that flips the state is the correct fix for the TOCTOU that Wave-5's generic `owned_method` (which relies on the public `lease()` path) leaves behind. It also lines up cleanly with `ForeignObjectLease`'s already-lease-deferred, lock-free decref.
+- **Separate `semantic_close_method` constructor** rather than overloading `owned_method`. It keeps the Wave-5 transport a straight bulk-move-with-a-lease and localizes the identity finalization logic to the one path that owns it.
+- **Threading the class cleanup policy from `class_method_emitter` into `(async_)python_interop_method_body`.** This is the right hook — HIR validation in `class_body_lowering.rs:614-640` already normalizes cleanup=async_close to exactly-one `Self.aclose` consuming coroutine, so codegen should drive selection off that class-level metadata rather than the method-name inference `python_interop_direct.rs:213-229` uses for sync `Self.close`.
+- **Reusing existing must-use/affine lowering** (`mod_context.rs:243-249` already maps `AsyncClose` to `CloseLike`) is correct; no new obligation kind needed.
+- **Keeping both public gates intact** matches the Wave-6 charter (`plans/issues/active/…m.d`:571-590) and the gate points at `python_interop.rs:364-366` and coroutine reservation. Wave 7 is the atomic-activation step.
+
+## Design changes required
+
+1. **Cancellation-vs-poison rule conflicts with the architecture doc.** Item 3 lists "cancellation/CancelledError" as a Closing→Poisoned trigger, and item 6 says "cancellation in flight waits Python finally then poisons." But `internal_docs/python_interop_protocol_architecture.md:104-106` states that if Python suppresses cancellation and returns or raises "the later return or exception wins normally." As written, a Sifr-initiated cancellation whose Python `finally` cleanly returns None would still poison, contradicting the doc. Fix by keying the terminal transition purely on the Python task's terminal outcome (clean `None` → Closed; any exception, including `CancelledError`, → Poisoned) and stop conflating "cancellation was requested" with "cancellation observed at terminal."
+
+2. **Pre-registration failure path is under-specified.** Once `begin_semantic_close` has flipped Open→Closing, any synchronous failure before the loop-thread setup callback runs (keyword validation, `ensure_started`, `reserve_submission` under `AsyncRuntimeStopping`, `call_soon_threadsafe`) must also finalize Closing→Poisoned before returning the error to the Sifr caller. Today's `submit_async_declaration` (`async_declaration.rs:21-48`) can fail on any of those paths; the design mentions "setup errors" but doesn't explicitly say the ObjectHandle must be poisoned on the synchronous submission-error branch. This has to be a documented, tested branch (a `Drop` guard on the request that poisons unless we handed ownership to a live submission is the natural implementation).
+
+3. **Emitter-selection guard should not be "consumes_receiver alone."** Item 4 says the emitter runs "on cleanup=AsyncClose." Please make that explicit and layered: (a) class cleanup=AsyncClose, (b) method kind=Coroutine, (c) target segments==`["Self","aclose"]`, (d) `consumes_receiver==true`, (e) params empty + return `Result[None, PythonError]`. The class-body check already enforces (a)+(c)+(d)+(e); the codegen should not rely on that check by accident. Otherwise a future non-cleanup consuming async method (or a bug that lets one through HIR) silently rewrites into the identity-finalizing emitter.
+
+4. **Maintainability under the 900-line cap.** `python_interop_async.rs` is already **826/900** lines. Adding the semantic constructor, cleanup-policy threading, and its tests will breach the cap. The plan should say up-front how the file is split (e.g., a `python_interop_async/` module with `function.rs`, `method.rs`, `semantic_close.rs`, shared helpers), rather than "add focused code" that later has to be reshuffled under `scripts/check_hir_maintainability_guardrails.py`. Same watch for `class_method_emitter.rs` (756 lines) when a new arg is threaded through.
+
+5. **Explicit "shutdown drains in-flight semantic closes" hook, not just "reuse cancel_registered_submissions."** Today `run_registered_async_cleanup()` in `shutdown_hooks.rs:21-29` is a no-op stub. The design says shutdown "cancels/drains an in-flight async close before loop join and leaves no task/thread." Please specify whether in-flight semantic closes ride the ordinary submission registry (which `shutdown()` in `async_runtime.rs:372-455` already cancels and drains) or need a separate, ordered slot ahead of the generic cancel step so that the loop still runs Python `finally` for close-tasks before other cancellations fire. The former is fine, but the design should state it and cover the ordering test (a stalled aclose blocks shutdown until Python `finally` runs, then the state transitions Closing→Poisoned via the terminal callback, and the loop is joined with zero tasks/threads left).
+
+6. **Owner-drop-without-close is a runtime, not just codegen, invariant.** Because `ForeignObject` derives `Clone` and runtime aliases can exist (e.g., through `PythonAsyncObject::owned` re-owning an object returned by convert_output), please state whether `Drop for ForeignObject` on the *last* Arc while state is still `Open` for an `AsyncClose` opaque is a runtime panic, a diagnostic, or a silent Poisoned (currently `ForeignObjectInner::drop` just drops the `Py<PyAny>` under GIL/pending-queue). Compile-time affine tracking should prevent Sifr code from getting here, but the runtime behavior when a bug slips through should be explicit rather than "the last Arc drops with `Open` intact."
+
+## Nice-to-haves (not blockers)
+
+- Runtime test list should include: (a) `begin_semantic_close` racing with a concurrent `lease()` on an aliased handle, (b) `begin_semantic_close` on already-Closing (must reject), (c) shutdown started **after** Open→Closing but **before** submission (poisons on the synchronous error, no task appears in the registry), (d) Python `aclose` returns a non-None awaitable value (conversion error → Poisoned), (e) task callback panic path also poisons (already exercised for typed decls; assert for this constructor).
+- Codegen tests should assert the emitter picks `semantic_close_method` **only** when all guard conditions hold (add a negative case with cleanup=AsyncClose but wrong method name or missing `own self` to confirm HIR would have rejected first, and the codegen path doesn't paper over it).
+
+VERDICT: CHANGES_REQUIRED
