+++ plans/reviews/active/python-interop-m7-wave6-async-close-implementation-review-round2.md
@@ -1,0 +1,1 @@
+VERDICT: SATISFIED
